+++ package.json
@@ -15,7 +15,6 @@
     "url": "https://github.com/gonzalovazquez/filmy-node.git"
   },
   "devDependencies": {
-    "mongojs": "^0.13.1",
     "express": "^4.8.3",
     "body-parser": "^1.6.3",
     "cors": "^2.4.1",
@@ -25,6 +24,6 @@
     "moment": "^2.8.1",
     "morgan": "^1.2.2",
     "request": "^2.40.0",
-    "bcryptjs": "^2.0.2"
+    "bcryptjs": "^0.7.12",
   }
 }
